infra/src/requirements.txt
aws-cdk.assets==1.32.2

aws-cdk.aws-cloudformation==1.32.2
aws-cdk.aws-codebuild==1.32.2
aws-cdk.aws-codecommit==1.32.2
aws-cdk.aws-codedeploy==1.32.2
aws-cdk.aws-codepipeline==1.32.2
aws-cdk.aws-codepipeline-actions==1.32.2
aws-cdk.aws_codestarconnections==1.32.2

aws-cdk.aws-iam==1.32.2
aws-cdk.aws-kms==1.32.2
aws-cdk.aws-lambda==1.32.2

aws-cdk.aws-s3==1.32.2
aws-cdk.aws-s3-assets==1.32.2
aws-cdk.aws-sns==1.32.2
aws-cdk.aws-sns-subscriptions==1.32.2

aws-cdk.aws-stepfunctions==1.32.2
aws-cdk.core==1.32.2

